requirements.txt
langchain==0.0.259 # 지금 안씀
langchain_community
pymysql
load_dotenv